little bit of navbar and hero work
