+++ app/layout.tsx
@@ -46,7 +46,6 @@ export default function RootLayout({ children }: RootLayoutProps) {
                 className={`antialiased min-h-screen bg-[#ffffff] dark:bg-slate-950 text-slate-900 dark:text-slate-50 ${inter.className}`}
             >
                 <div>
-                    <Navbar />
                     <main>{children}</main>
                     <Footer />
                 </div>
