+++ app/page.tsx
@@ -3,11 +3,13 @@ import PageHero from "@/components/PageHero"
 import ProjectsSection from "@/components/ProjectsSection"
 import WorkedWith from "@/components/WorkedWith"
 import ResearchAreasParent from "@/components/ResearchAreasParent"
+import Navbar from "@/components/Navbar"
 
 export default function Home() {
     return (
         <>
-            <PageHero title="Human Aware Planning and Interaction Lab" />
+            <Navbar />
+            <PageHero title="HUMAN-AWARE PLANNING AND INTERACTION LAB" />
             <div className="w-full h-8 bg-[#c69b3d]"></div>
             <ResearchAreasParent />
             <MeetLab />
